resumeai/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
scikit-learn
sentence-transformers
spacy
numpy
pandas
anthropic
openai
python-dotenv
aiofiles
httpx
joblib
Pillow
pydantic[email]
google-generativeai
